optimize xapian index, don't use doc id and dedup on search
using the doc id slows down indexing with xapian, as it's like a primary key.

instead, we already know that we won't duplicate inserts of versions as we keep
track of the last inserted. And just in case, we can use the collapse feature of
xapian to remove any duplicates as search time, which is fast.